Minor commit, next "TODO" could break the library so commit is to get the most up-to-date working version of the library and provide backup if something goes wrong.
Consolidated functions in cmspack. Previously the callbacks would simply call the actual function to perform the operation. Now the callbacks are the functions. This reduces function calls, uses a (very) little less memory, and should be faster when callbacks are used many times.
Tracing of basic endian operations for functions that have the option to reverse endian. Unneeded functions removed and "final" function is called directly instead of through overloaded function.